refactor(app): fix ESLint script and fix/disable rules
