+++ src/__mocks__/bookmarks.json
@@ -1,5 +1,17 @@
 [
-  { "title": "Bookmark test 1", "url": "http://bookmarktest1.com", "id": 1 },
-  { "title": "Bookmark test 2", "url": "http://bookmarktest2.com", "id": 2 },
-  { "title": "Bookmark test 3", "url": "http://bookmarktest3.com", "id": 3 }
+  {
+    "title": "Bookmark test 1",
+    "url": "http://bookmarktest1.com",
+    "id": 1
+  },
+  {
+    "title": "Bookmark test 2",
+    "url": "http://bookmarktest2.com",
+    "id": 2
+  },
+  {
+    "title": "Bookmark test 3",
+    "url": "http://bookmarktest3.com",
+    "id": 3
+  }
 ]
