+++ src/components/01-Atoms/Icon/Icon.jsx
@@ -1,3 +1,5 @@
+/* eslint-disable react/prop-types */
+
 import React from "react";
 import { Menu, Plus as Add, LogOut } from "react-feather";
