+++ src/components/02-Molecules/IconButton/IconButton.jsx
@@ -1,3 +1,5 @@
+/* eslint-disable react/prop-types */
+
 import React from "react";
 import styled from "styled-components";
