+++ src/components/03-Organisms/Footer/Footer.jsx
@@ -1,3 +1,5 @@
+/* eslint-disable react/prop-types */
+
 import React from "react";
 
 export default function Footer({ isUserSignedIn }) {
