+++ src/components/04-Ecosystems/MainNavbar/MainNavbar.jsx
@@ -1,7 +1,9 @@
+/* eslint-disable react/prop-types */
+
 import React from "react";
 
 import IconButton from "../../02-Molecules/IconButton";
-import { ReactComponent as Logo } from "../../assets/images/agenda.svg";
+import { ReactComponent as Logo } from "../../../assets/images/agenda.svg";
 
 export default function MainNavbar({ handleSignOut, handleShowAddBookmark }) {
   return (
